Updates webpack & optimizes dependencies & improves settings code.
