+++ package.json
@@ -5,44 +5,41 @@
 	"author": "Henning Dieterichs <henning.dieterichs@live.de>",
 	"license": "UNLICENSED",
 	"scripts": {
-		"dev": "webpack --watch --mode development --use-cdn-for-monaco",
-		"dev-without-cdn": "webpack --watch --mode dev",
+		"dev": "webpack --watch --mode development --env=use-cdn-for-monaco=true",
+		"dev-without-cdn": "webpack --watch --mode development",
 		"build": "yarn build-js && yarn bundle",
 		"build-js": "webpack --mode production",
 		"bundle": "yarn shx rm -rf ./dist/extension.zip && yarn jszip-cli add -o ./dist/extension.zip -i *.map dist manifest.json logo.drawio.png"
 	},
 	"devDependencies": {
 		"@ffflorian/jszip-cli": "^3.1.0",
-		"@types/chrome": "^0.0.114",
-		"@types/classnames": "^2.2.9",
+		"@types/chrome": "^0.0.133",
 		"@types/html-webpack-plugin": "^3.2.2",
-		"@types/react": "^16.9.22",
-		"@types/react-dom": "^16.9.5",
-		"@types/react-measure": "^2.0.6",
+		"@types/mini-css-extract-plugin": "^1.2.2",
 		"@types/webpack": "^4.41.6",
 		"clean-webpack-plugin": "^3.0.0",
-		"css-loader": "^3.4.2",
-		"file-loader": "^5.1.0",
-		"fork-ts-checker-webpack-plugin": "^4.0.4",
-		"html-webpack-plugin": "^3.2.0",
-		"monaco-editor-webpack-plugin": "^1.9.0",
+		"css-loader": "^5.1.3",
+		"file-loader": "^6.2.0",
+		"fork-ts-checker-webpack-plugin": "^6.2.0",
+		"html-webpack-plugin": "^5.3.1",
+		"mini-css-extract-plugin": "^1.3.3",
+		"monaco-editor-webpack-plugin": "^3.0.1",
 		"prettier": "^2.2.1",
 		"raw-loader": "^4.0.0",
 		"sass": "^1.25.0",
-		"sass-loader": "^8.0.2",
+		"sass-loader": "^11.0.1",
 		"shx": "^0.3.2",
-		"style-loader": "^1.1.3",
-		"ts-loader": "^6.2.1",
-		"ts-node": "^8.6.2",
+		"style-loader": "^2.0.0",
+		"ts-loader": "^8.0.18",
+		"ts-node": "^9.1.1",
 		"typescript": "^4.1.3",
-		"webpack": "^4.41.6",
-		"webpack-cli": "^3.3.11",
-		"@types/mini-css-extract-plugin": "^1.2.2",
-		"mini-css-extract-plugin": "^1.3.3"
+		"webpack": "^5.27.0",
+		"webpack-cli": "^4.5.0"
 	},
 	"dependencies": {
-		"@hediet/monaco-editor-react": "^0.1.1",
+		"@types/copy-webpack-plugin": "^6.4.0",
+		"copy-webpack-plugin": "^8.0.0",
 		"gemoji": "^6.1.0",
-		"monaco-editor": "^0.20.0"
+		"monaco-editor": "^0.23.0"
 	}
 }
